+++ .eslintrc.json
@@ -1,64 +1,82 @@
 {
   "extends": [
     "eslint:recommended",
+    "plugin:@typescript-eslint/recommended", // Uses the recommended rules from the @typescript-eslint/eslint-plugin
+    "plugin:lit/recommended",
     "google",
-    "plugin:prettier/recommended"
-    // Enables eslint-plugin-prettier and eslint-config-prettier.
+    "plugin:prettier/recommended" // Enables eslint-plugin-prettier and eslint-config-prettier.
+    //This will display prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
   ],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
-    "ecmaVersion": 6,
+    "ecmaVersion": 2018,
     "sourceType": "module"
   },
   "env": {
-    "browser": true
+    "browser": true,
+    "node": true,
+    "es6": true
   },
-  "plugins": ["html", "prettier"],
+  "plugins": ["@typescript-eslint", "lit", "prettier"],
   "rules": {
-    "padded-blocks": "off",
+    "prettier/prettier": "error",
+    "lit/attribute-value-entities": "off",
+    "lit/no-legacy-template-syntax": "off",
     "linebreak-style": "off",
-    "brace-style": "off",
-    "new-cap": [
-      "error",
-      {
-        "capIsNewExceptions": ["Polymer", "ListItemUtils", "MissingOptions", "CommonFunctionality"],
-        "capIsNewExceptionPattern": "^Etools.."
-      }
-    ],
-    "no-var": "off",
-    "require-jsdoc": "off",
-    "valid-jsdoc": "off",
-    "comma-dangle": ["error", "never"],
     "no-irregular-whitespace": [
       "error",
       {
         "skipTemplates": true
       }
     ],
+    "@typescript-eslint/no-object-literal-type-assertion": "off",
+    "padded-blocks": "off",
+    "brace-style": "off",
+    "new-cap": [
+      "error",
+      {
+        "capIsNewExceptions": ["GestureEventListeners"],
+        "capIsNewExceptionPattern": "^Etools..|..Mixin$"
+      }
+    ],
+    "no-var": "off",
+    "require-jsdoc": "off",
+    "valid-jsdoc": "off",
+    "comma-dangle": ["error", "never"],
     "max-len": [
       "error",
       {
-        "code": 120
+        "code": 120,
+        "ignoreUrls": true
       }
     ],
-    "camelcase": [
+    "prefer-promise-reject-errors": "off",
+    "camelcase": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off",
+    "@typescript-eslint/explicit-member-accessibility": "off",
+    "@typescript-eslint/camelcase": "off",
+    "@typescript-eslint/explicit-function-return-type": "off",
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-non-null-assertion": "off",
+    "@typescript-eslint/ban-ts-comment": "off",
+    "@typescript-eslint/no-non-null-asserted-optional-chain": "off",
+    "@typescript-eslint/ban-types": ["warn"],
+    "@typescript-eslint/no-use-before-define": [
       "error",
       {
-        "properties": "never"
+        "functions": false,
+        "classes": true,
+        "variables": true
       }
     ],
-    "arrow-parens": [
-      2,
-      "always",
+    "@typescript-eslint/no-unused-vars": [
+      "error",
       {
-        "requireForBlockBody": true
+        "argsIgnorePattern": "^_"
       }
     ]
   },
   "globals": {
-    "moment": true,
-    "Promise": true,
-    "Polymer": true,
-    "EtoolsLogsMixin": true,
-    "EsmmMixins": true
+    "dayjs": true
   }
 }
