--- assets/translations.js
@@ -1,31 +1,0 @@
-const translations = {
-  en: {
-    ROWS_PER_PAGE: 'Rows per page:',
-    OF: 'of'
-  },
-  fr: {
-    ROWS_PER_PAGE: 'Lignes par page:',
-    OF: 'sur'
-  },
-  ro: {
-    ROWS_PER_PAGE: 'Rânduri pe pagină:',
-    OF: 'din'
-  },
-  pt: {
-    ROWS_PER_PAGE: 'Linhas por página:',
-    OF: 'de'
-  },
-  ru: {
-    ROWS_PER_PAGE: 'Рядов на странице:',
-    OF: 'из'
-  },
-  es: {
-    ROWS_PER_PAGE: 'Filas por página:',
-    OF: 'de'
-  },
-  ar: {
-    ROWS_PER_PAGE: 'دد الصفوف في الصفحة:', // The browser moves the colon to the left
-    OF: 'من'
-  }
-};
-export default translations;
