+++ package.json
@@ -3,8 +3,31 @@
   "description": "Etools table",
   "version": "2.0.1",
   "homepage": "https://github.com/unicef-polymer/etools-table",
-  "main": "etools-table.js",
-  "module": "etools-table.js",
+  "main": "./src/etools-table.js",
+  "module": "./src/etools-table.js",
+  "files": [
+    "src"
+  ],
+  "exports": {
+    "./etools-table": {
+      "types": "./src/etools-table.d.ts",
+      "default": "./src/etools-table.js"
+    },
+    "./pagination/etools-pagination": {
+      "types": "./src/pagination/etools-pagination.d.ts",
+      "default": "./src/pagination/etools-pagination.js"
+    }
+  },
+  "typesVersions": {
+    "*": {
+      "pagination/etools-pagination": [
+        "./src/pagination/etools-pagination.d.ts"
+      ],
+      "etools-table": [
+        "./src/etools-table.d.ts"
+      ]
+    }
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/unicef-polymer/etools-table.git"
@@ -15,29 +38,33 @@
     "@polymer/paper-checkbox": "^3.1.0",
     "@polymer/paper-icon-button": "^3.0.2",
     "@polymer/polymer": "^3.4.1",
-    "lit-element": "^2.4.0",
+    "@shoelace-style/shoelace": "^2.5.2",
+    "lit-element": "^2.5.1",
+    "lit-html": "^2.7.5",
     "lodash-es": "^4.17.21"
   },
   "devDependencies": {
-    "@webcomponents/webcomponentsjs": "^2.6.0",
-    "acorn": "^8.0.5",
-    "diff": "^5.0.0",
+    "@types/lodash-es": "^4.17.8",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "@typescript-eslint/typescript-estree": "^5.7.0",
     "eslint": "^7.20.0",
     "eslint-config-google": "^0.13.0",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-html": "^6.1.1",
     "eslint-plugin-prettier": "^3.2.0",
-    "growl": ">=1.10.0",
-    "minimist": ">=1.2.6",
-    "minimatch": ">=3.1.2",
+    "glob-parent": ">=5.1.2",
+    "minimist": ">=0.2.1",
     "prettier": "^2.3.2",
-    "wct-browser-legacy": "^1.0.2",
-    "glob-parent": ">=5.1.2"
+    "typescript": "^4.5.4"
   },
   "scripts": {
-    "lint": "eslint . --ext js,html",
-    "prepublishOnly": "npm run lint",
-    "publish": "git push --tags && npm publish"
+    "tsToJs": "tsc",
+    "watch": "tsc --watch",
+    "lint": "eslint . --ext js,html --ignore-path .gitignore",
+    "format": "eslint --ext .ts . --fix --ignore-path .gitignore",
+    "publishPackage": "npm run tsToJs && npm publish",
+    "prepublishOnly": "npm run tsToJs"
   },
   "publishConfig": {
     "access": "public"
