+++ src/assets/translations.d.ts
@@ -1,0 +1,31 @@
+declare const translations: {
+    en: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+    fr: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+    ro: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+    pt: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+    ru: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+    es: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+    ar: {
+        ROWS_PER_PAGE: string;
+        OF: string;
+    };
+};
+export default translations;
