+++ src/styles/grid-layout-styles.d.ts
@@ -1,0 +1,1 @@
+export declare const gridLayoutStylesLit: import("lit-element").CSSResult;
