+++ src/styles/pagination-style.d.ts
@@ -1,0 +1,1 @@
+export declare const etoolsPaginationStyles: import("lit-element").CSSResult;
