+++ src/styles/table-responsive-styles.d.ts
@@ -1,0 +1,1 @@
+export declare const etoolsTableResponsiveStyles: import("lit-element").CSSResult;
