+++ src/styles/table-styles.d.ts
@@ -1,0 +1,1 @@
+export declare const etoolsTableStyles: import("lit-element").CSSResult;
