+++ src/utils/date-utility.d.ts
@@ -1,0 +1,4 @@
+export declare function isValidDate(date: any): boolean;
+export declare function prettyDate(dateString: string, format: string, placeholder?: string): any;
+export declare function convertDate(dateString: string, noZTimezoneOffset?: string): Date | null;
+export declare function formatDate(date: any, format: string): any;
