+++ src/utils/fire-custom-event.d.ts
@@ -1,0 +1,1 @@
+export declare const fireEvent: (el: any, eventName: string, eventDetail: any) => void;
