+++ src/utils/fire-custom-event.js
@@ -1,0 +1,7 @@
+export const fireEvent = (el, eventName, eventDetail) => {
+    el.dispatchEvent(new CustomEvent(eventName, {
+        detail: eventDetail,
+        bubbles: true,
+        composed: true
+    }));
+};
