+++ src/utils/translate.d.ts
@@ -1,0 +1,1 @@
+export declare function getTranslation(lang: string, key: string): any;
