+++ src/utils/utils.d.ts
@@ -1,0 +1,2 @@
+export declare const fireEvent: (el: any, eventName: string, eventDetail: any) => void;
+export declare const toggleAttributeValue: (el: any, attrName: string, attrVal1: any, attrVal2: any) => void;
