--- utils/fire-custom-event.js
@@ -1,9 +1,0 @@
-export const fireEvent = (el, eventName, eventDetail) => {
-  el.dispatchEvent(
-    new CustomEvent(eventName, {
-      detail: eventDetail,
-      bubbles: true,
-      composed: true
-    })
-  );
-};
