--- utils/translate.js
@@ -1,9 +1,0 @@
-import translations from '../assets/translations';
-
-export function getTranslation(lang, key) {
-  try {
-    return translations[lang][key];
-  } catch (error) {
-    return translations.en[key];
-  }
-}
